test: separate recommendation behavior tests
